Section 3.3.1: Top-Down Parser (leftmost, oracular).
This is not meant to work with the classic grammar because the code
cannot know which expansion to use, so it loops indefinitely, repeatedly
left recursing.

It does work with the right-recursive variant of the grammar.

This code is far more complicated than the pseudo code suggests, mostly
to make sure that backtracking works correctly. I would be glad if
someone showed me how to simplify this